The changes that have been made address the following issue.
When a computer is inventoried using a local standalone audit 
on one subnet, it stays on that subnet in the ipdiscover page. 
If the computer is moved to a different subnet, it still does 
not show up on the new subnet at all, even though it exists in 
the netmap table on the new subnet. The changes in the 
ipdiscover file modify it so that the inventoried subnet and 
the scanned subnet are compared for each mac address. If the 
subnets are not equal, the computer is now listed as 
non-inventoried. This is done through only modifying sql 
statements.

    The statements that have been modified are as follows:
    The two statements that give the total have been modified 
	to compare subnets.
    The statement that selects non-inventoried for each subnet 
	has been modified to not exclude all inventoried mac 
	address, but only inventoried mac address on that subnet.
    The statement that lists the non-inventoried computers has 
	been modified the same way.

All modifications have comments before them.