fix this bug: https://bugs.launchpad.net/bugs/585077
When opening the Date Time Picker in the administrative data for an inventoried computer, the year comes up as 2005. I tracked down the problem to the file /usr/share/ocsinventory-reports/ocsreports/js/datetimepicker.js on line 108 the date is being reset to 2005. I fixed this problem in my own installation by adding a line 109 that reads "Cal.Year = dtToday.getFullYear();", however it would be nice if it were fixed in subsequent versions. Thanks,

tx to Natrinicle