Changed what's set in the FROM field when getting the list of installed software on Debian.
The FROM field used to be statically set to 'deb', but changing it to the {Homepage} field
of dpk-query will help into finding the upstream project developing the software.

Knowing the upstream project of a software might help in precisely identifying a software,
which might help in, for example, finding a corresponding CPE entry in NIST's CPE dictionary.

It should be 'a good thing' to have the same change for rpm packages,... but I have no such platform
to test the change here atm.