Addendum:
1) The Full-text search doesn't work well yet. I think I need to tune how the fulltext index
is created or how the fulltext search is made
2) This PoC breaks the search functionality in the software table of a single agent.

===========================

First proof of concept of full-text searchable columns.

This PoC is able to do a full-text search when the user searches in the "All software" page
(visu_all_soft), while not breaking anything else AFAIK.

This PoC automatically detects when a column is full-text indexed and modify the search query
accordingly.

To create a full-text index against the "PUBLISHER" column in the "ocsweb.softwares" table,
 use the following command:

CREATE FULLTEXT INDEX publisher_ft_idx ON softwares (PUBLISHER);