Addendum:
This commit still has the following issues (at least):

1) when looking for "Oracle Java", any of the word may match for a result to be returned.
That would probably be better if both words had to match for a result to be returned.
I believe this can be achieved quite easily with the "IN BOOLEAN MODE" search modifier.
2) The "(MATCH (column1) AGAINST (expr) AND MATCH (column2) AGAINST (expr))" clause can be simplified
as we can simply do "(MATCH (column1,column2) AGAINST (expr)"
3) This PoC still can only search in the "softwares" table.

============================================================

This commit updates the full-text search functionality started in commit
581ba50287a66c78f5df8887a2beadcc4f8245d8.

The previous implementation didn't work because the resulting query was mixing "WHERE" and "HAVING" clause.

Example:
  SELECT * FROM sofwares WHERE (MATCH (NAME) AGAINST ('Atlassian')) GROUP BY NAME HAVING (PUBLISHER LIKE '%%Atlassian%%');

The problem is that the WHERE clause and the HAVING clause where AND'ed. So, in the above example, both the NAME and the
PUBLISHER columns match.

This commit solved this issue.

This commit also doesn't break the search functionality in the software table of a single agent anymore.